backend/pyproject.toml
[tool.poetry]
name = "aerae-accelerator-backend"
version = "0.1.0"
description = "Backend API for the AERAE Accelerator platform"
authors = ["AERAE Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.34.0", extras = ["standard"] }
pydantic-settings = "^2.7.0"
sqlmodel = "^0.0.22"
chromadb = "^0.6.0"
google-genai = "^1.0.0"
openai = "^1.60.0"
google-generativeai = "^0.8.0"
gitpython = "^3.1.0"
requests = "^2.32.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
ruff = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120
